Feature/bngp 3891 check and submit and application data (#508)
* moving applicant info to common view and route helper

* update missed from merge

* updates from 3622

* manual merge change

* manual merge from 3622

* taking out dummy test

* reverting changes to test file
