+++ packages/webapp/src/views/land/check-and-submit.html
@@ -12,6 +12,11 @@
         <h1 class="govuk-heading-xl">
           {{ pageHeading }}
         </h1>
+        
+        <!--Applicant information-->
+        <h2 class="govuk-heading-l">Applicant information</h2>
+        {% include "common/applicant-information.html" %}
+
         <!--Land boundary-->
         <h2 class="govuk-heading-l">Land boundary</h2>
